+++ test/XRoadLib.Tests/Protocols/Description/ProducerDefinitionTest.cs
@@ -1,18 +1,14 @@
 ﻿using System;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Xml.Linq;
 using XRoadLib.Extensions;
-using XRoadLib.Tests.Contract;
 using Xunit;
 
 namespace XRoadLib.Tests.Protocols.Description
 {
     public class ProducerDefinitionTest
     {
-        private static readonly Assembly contractAssembly = typeof(Class1).GetTypeInfo().Assembly;
-
         private static readonly Func<string, XName> wsdl = x => XName.Get(x, NamespaceConstants.WSDL);
         private static readonly Func<string, XName> soap = x => XName.Get(x, NamespaceConstants.SOAP);
         private static readonly Func<string, XName> xroad = x => XName.Get(x, NamespaceConstants.XROAD);
@@ -30,9 +26,9 @@ namespace XRoadLib.Tests.Protocols.Description
             Assert.NotNull(address);
             Assert.True(address.IsEmpty);
             Assert.Single(address.Attributes());
-            Assert.Equal("test-producer", address.Attribute("producer").Value);
+            Assert.Equal("test-producer", address.Attribute("producer")?.Value);
 
-            Assert.Equal("http://TURVASERVER/cgi-bin/consumer_proxy", port.Elements(soap("address")).Single().Attribute("location").Value);
+            Assert.Equal("http://TURVASERVER/cgi-bin/consumer_proxy", port.Elements(soap("address")).Single().Attribute("location")?.Value);
         }
 
         [Fact]
@@ -45,9 +41,9 @@ namespace XRoadLib.Tests.Protocols.Description
             Assert.NotNull(address);
             Assert.True(address.IsEmpty);
             Assert.Single(address.Attributes());
-            Assert.Equal("test-producer", address.Attribute("producer").Value);
+            Assert.Equal("test-producer", address.Attribute("producer")?.Value);
 
-            Assert.Equal("http://TURVASERVER/cgi-bin/consumer_proxy", port.Elements(soap("address")).Single().Attribute("location").Value);
+            Assert.Equal("http://TURVASERVER/cgi-bin/consumer_proxy", port.Elements(soap("address")).Single().Attribute("location")?.Value);
         }
 
         [Fact]
@@ -56,7 +52,7 @@ namespace XRoadLib.Tests.Protocols.Description
             var url = "http://TURVASERVER/cgi-bin/consumer_proxy";
             var doc = GetDocument(Globals.ServiceManager31, 1);
             var port = GetPort(doc, xroad);
-            Assert.Equal(url, port.Elements(soap("address")).Single().Attribute("location").Value);
+            Assert.Equal(url, port.Elements(soap("address")).Single().Attribute("location")?.Value);
         }
 
         [Fact]
@@ -76,17 +72,17 @@ namespace XRoadLib.Tests.Protocols.Description
                               x =>
                               {
                                   Assert.Equal("XRoadLib test producer", x.Value);
-                                  Assert.Equal("en", x.Attribute(xml("lang")).Value);
+                                  Assert.Equal("en", x.Attribute(xml("lang"))?.Value);
                               },
                               x =>
                               {
                                   Assert.Equal("XRoadLib test andmekogu", x.Value);
-                                  Assert.Equal("et", x.Attribute(xml("lang")).Value);
+                                  Assert.Equal("et", x.Attribute(xml("lang"))?.Value);
                               },
                               x =>
                               {
                                   Assert.Equal("Portugalikeelne loba ...", x.Value);
-                                  Assert.Equal("pt", x.Attribute(xml("lang")).Value);
+                                  Assert.Equal("pt", x.Attribute(xml("lang"))?.Value);
                               });
 
             var noCode = groupedByAttributes.Where(x => !x.Key).SelectMany(x => x).ToList();
@@ -103,7 +99,7 @@ namespace XRoadLib.Tests.Protocols.Description
             var titleElements = port.Elements(xtee("title")).ToList();
             Assert.Equal(4, titleElements.Count);
 
-            var groupedByAttributes = titleElements.GroupBy(x => x.Attributes().Any());
+            var groupedByAttributes = titleElements.GroupBy(x => x.Attributes().Any()).ToList();
 
             var yesCode = groupedByAttributes.Where(x => x.Key).SelectMany(x => x).ToList();
             Assert.Equal(3, yesCode.Count);
@@ -111,17 +107,17 @@ namespace XRoadLib.Tests.Protocols.Description
                               x =>
                               {
                                   Assert.Equal("XRoadLib test producer", x.Value);
-                                  Assert.Equal("en", x.Attribute(xml("lang")).Value);
+                                  Assert.Equal("en", x.Attribute(xml("lang"))?.Value);
                               },
                               x =>
                               {
                                   Assert.Equal("XRoadLib test andmekogu", x.Value);
-                                  Assert.Equal("et", x.Attribute(xml("lang")).Value);
+                                  Assert.Equal("et", x.Attribute(xml("lang"))?.Value);
                               },
                               x =>
                               {
                                   Assert.Equal("Portugalikeelne loba ...", x.Value);
-                                  Assert.Equal("pt", x.Attribute(xml("lang")).Value);
+                                  Assert.Equal("pt", x.Attribute(xml("lang"))?.Value);
                               });
 
             var noCode = groupedByAttributes.Where(x => !x.Key).SelectMany(x => x).ToList();
@@ -135,18 +131,18 @@ namespace XRoadLib.Tests.Protocols.Description
             var doc = GetDocument(Globals.ServiceManager31, 1u);
             var definitions = doc.Elements(wsdl("definitions")).Single();
             var types = definitions.Elements(wsdl("types")).Single();
-            var schema = types.Elements(xsd("schema")).Single();
-            Assert.DoesNotContain(schema.Elements(xsd("complexType")), e => e.Attribute("name").Value == "AnonymousType");
+            var schema = types.Elements(xsd("schema")).Single(x => x.Attribute("targetNamespace")?.Value == Globals.ServiceManager31.ProducerNamespace);
+            Assert.DoesNotContain(schema.Elements(xsd("complexType")), e => e.Attribute("name")?.Value == "AnonymousType");
 
-            var containerType = schema.Elements(xsd("complexType")).Single(e => e.Attribute("name").Value == "ContainerType");
+            var containerType = schema.Elements(xsd("complexType")).Single(e => e.Attribute("name")?.Value == "ContainerType");
             var containerTypeParticle = containerType.Elements().Single();
             Assert.Equal(xsd("sequence"), containerTypeParticle.Name);
             Assert.Equal(2, containerTypeParticle.Elements().Count());
 
-            var knownProperty = containerTypeParticle.Elements(xsd("element")).Single(e => e.Attribute("name").Value == "KnownProperty");
-            Assert.Equal($"{knownProperty.GetPrefixOfNamespace(NamespaceConstants.XSD)}:string", knownProperty.Attribute("type").Value);
+            var knownProperty = containerTypeParticle.Elements(xsd("element")).Single(e => e.Attribute("name")?.Value == "KnownProperty");
+            Assert.Equal($"{knownProperty.GetPrefixOfNamespace(NamespaceConstants.XSD)}:string", knownProperty.Attribute("type")?.Value);
 
-            var anonymousProperty = containerTypeParticle.Elements(xsd("element")).Single(e => e.Attribute("name").Value == "AnonymousProperty");
+            var anonymousProperty = containerTypeParticle.Elements(xsd("element")).Single(e => e.Attribute("name")?.Value == "AnonymousProperty");
             Assert.Null(anonymousProperty.Attribute("type"));
 
             var anonymousType = anonymousProperty.Elements().Single();
@@ -158,9 +154,9 @@ namespace XRoadLib.Tests.Protocols.Description
             Assert.Equal(3, anonymousSequence.Elements().Count());
 
             Assert.Collection(anonymousSequence.Elements(),
-                              x => Assert.Equal("Property1", x.Attribute("name").Value),
-                              x => Assert.Equal("Property2", x.Attribute("name").Value),
-                              x => Assert.Equal("Property3", x.Attribute("name").Value));
+                              x => Assert.Equal("Property1", x.Attribute("name")?.Value),
+                              x => Assert.Equal("Property2", x.Attribute("name")?.Value),
+                              x => Assert.Equal("Property3", x.Attribute("name")?.Value));
         }
 
         private XElement GetPort(XContainer doc, Func<string, XName> xrdns)
@@ -171,14 +167,14 @@ namespace XRoadLib.Tests.Protocols.Description
             var service = root.Elements(wsdl("service")).SingleOrDefault();
             Assert.NotNull(service);
             Assert.Single(service.Attributes());
-            Assert.Equal("TestService", service.Attribute("name").Value);
+            Assert.Equal("TestService", service.Attribute("name")?.Value);
             Assert.Single(service.Elements());
 
             var port = service.Elements(wsdl("port")).SingleOrDefault();
             Assert.NotNull(port);
             Assert.Equal(2, port.Attributes().Count());
-            Assert.Equal("TestPort", port.Attribute("name").Value);
-            Assert.Equal("tns:TestBinding", port.Attribute("binding").Value);
+            Assert.Equal("TestPort", port.Attribute("name")?.Value);
+            Assert.Equal("tns:TestBinding", port.Attribute("binding")?.Value);
 
             Assert.DoesNotContain(port.Elements(), e => e.Name != soap("address") && e.Name != xrdns("address") && e.Name != xrdns("title"));